Feature/chathurika

+++ backend/controllers/customerController.js
@@ -2,10 +2,18 @@ const asyncHandler = require("express-async-handler");
 const Customer = require("../models/customerModel");
 const generateToken = require("../utils/generateToken");
 const bcrypt = require("bcryptjs");
+const crypto = require("crypto");
+
+const CUSTOMER_SESSIONS = new Map();
 
 // register  customer profile
 const registerCustomer = asyncHandler(async (req, res) => {
 	const { firstName, lastName, telephone, address, gender, country, email, password, pic } = req.body;
+	const user = CUSTOMER_SESSIONS.get(req.cookies.sessionId);
+	if (user == null) {
+		res.sendStatus(401);
+		return;
+	}
 
 	const customerExists = await Customer.findOne({ email });
 	if (customerExists) {
@@ -67,6 +75,21 @@ const authCustomer = asyncHandler(async (req, res) => {
 		res.status(400);
 		throw new Error("Invalid Email or Password");
 	} else {
+		const sessionId = crypto.randomUUID();
+		const csrfToken = crypto.randomUUID();
+
+		const id = customer._id;
+		CUSTOMER_SESSIONS.set(sessionId, { id, csrfToken });
+		const expirationDate = new Date();
+		expirationDate.setDate(expirationDate.getDate() + 2);
+
+		//Set the cookie
+		res.cookie("sessionId", sessionId, {
+			httpOnly: false,
+			withCredentials: true,
+			expires: expirationDate,
+		});
+
 		res.status(201).json({
 			_id: customer._id,
 			firstName: customer.firstName,
@@ -79,6 +102,7 @@ const authCustomer = asyncHandler(async (req, res) => {
 			pic: customer.pic,
 			regDate: customer.regDate,
 			token: generateToken(customer._id),
+			csrfToken,
 		});
 	}
 });
@@ -116,6 +140,11 @@ const getCustomerProfileById = asyncHandler(async (req, res) => {
 //update customer profile by customer
 const updateCustomerProfile = asyncHandler(async (req, res) => {
 	const customer = await Customer.findById(req.customer._id);
+	const user = CUSTOMER_SESSIONS.get(req.cookies.sessionId);
+	if (user == null || user.csrfToken !== req.body.csrfToken) {
+		res.sendStatus(401);
+		return;
+	}
 
 	if (customer) {
 		customer.firstName = req.body.firstName || customer.firstName;
@@ -193,6 +222,12 @@ const updateCustomerProfileById = asyncHandler(async (req, res) => {
 const deleteCustomerProfile = asyncHandler(async (req, res) => {
 	const customer = await Customer.findById(req.customer._id);
 
+	const user = CUSTOMER_.get(req.cookies.sessionId);
+	if (user == null) {
+		res.sendStatus(401);
+		return;
+	}
+
 	if (customer) {
 		await customer.deleteOne();
 		res.json({ message: "Customer Removed !" });
@@ -215,7 +250,25 @@ const deleteCustomerProfileById = asyncHandler(async (req, res) => {
 	}
 });
 
+// create csrf token
+const getCSRF = asyncHandler(async (req, res) => {
+	const sessionId = req.cookies.sessionId;
+	// Check if the session exists in CUSTOMER_SESSIONS
+	if (req.cookies.sessionId) {
+		const newCsrfToken = crypto.randomUUID();
+
+		// Assign the new CSRF token to the session object
+		const session = CUSTOMER_SESSIONS.get(req.cookies.sessionId);
+		session.csrfToken = newCsrfToken;
+
+		res.json({ newCsrfToken });
+	} else {
+		res.status(400).json({ message: "Session not found" });
+	}
+});
+
 module.exports = {
+	CUSTOMER_SESSIONS,
 	registerCustomer,
 	authCustomer,
 	getCustomers,
@@ -225,4 +278,5 @@ module.exports = {
 	updateCustomerProfileById,
 	deleteCustomerProfile,
 	deleteCustomerProfileById,
+	getCSRF,
 };
